Resolves Issue 1 - Change Local State to Context
Added TrainingContext and moved all local state from TrainingDay.js to context file
