+++ frontend/src/index.js
@@ -6,12 +6,15 @@ import { store } from "./app/store";
 import { Provider } from "react-redux";
 import * as serviceWorker from "./serviceWorker";
 import "./App.scss";
+import { StateContext } from "./features/Context/TrainingContext";
 
 ReactDOM.render(
   <React.StrictMode>
-    <Provider store={store}>
-      <App />
-    </Provider>
+    <StateContext>
+      <Provider store={store}>
+        <App />
+      </Provider>
+    </StateContext>
   </React.StrictMode>,
   document.getElementById("root")
 );
